refactor: home export and FFI unit tests
Move export.rs tests into export/tests.rs and the C-ABI coverage out of
lib.rs into gigastt-ffi/src/tests.rs. Public export and FFI surfaces are
unchanged.
